feat: add getInventoryListFromAFilter helper
