+++ dashboard/components/inventory/hooks/useInventory/useInventory.tsx
@@ -13,6 +13,7 @@ import {
   View
 } from './types/useInventoryTypes';
 import getInventoryListAndStats from './utils/getInventoryListAndStats';
+import getInventoryListFromAFilter from './utils/getInventoryListFromAFilter';
 import parseURLParams from './utils/parseURLParams';
 
 function useInventory() {
@@ -97,77 +98,6 @@ function useInventory() {
         }
       }
     });
-  }
-
-  /** Fetch inventory from a filter.
-   * - Inventory list will be stored in the state: searchedInventory
-   */
-  function getInventoryListFromAFilter() {
-    if (
-      router.query &&
-      Object.keys(router.query).length > 0 &&
-      !router.query.view
-    ) {
-      if (
-        Object.keys(router.query)[0].split(':').length <= 1 &&
-        !router.query.view
-      ) {
-        setTimeout(() => router.push('/'), 5000);
-        return setToast({
-          hasError: true,
-          title: `Invalid URL params`,
-          message: `There was an error processing the page. Redirecting back to home...`
-        });
-      }
-      setSearchedLoading(true);
-      setStatsLoading(true);
-
-      const newFilters: InventoryFilterData[] = Object.keys(router.query).map(
-        param => parseURLParams(param as string, 'fetch')
-      );
-      const newFiltersToDisplay: InventoryFilterData[] = Object.keys(
-        router.query
-      ).map(param => parseURLParams(param as string, 'display'));
-
-      const payloadJson = JSON.stringify(newFilters);
-
-      settingsService.getFilteredInventoryStats(payloadJson).then(res => {
-        if (res === Error) {
-          setError(true);
-          setStatsLoading(false);
-        } else {
-          setInventoryStats(res);
-          setStatsLoading(false);
-        }
-      });
-
-      settingsService
-        .getInventory(`?limit=${batchSize}&skip=0`, payloadJson)
-        .then(res => {
-          if (res.error) {
-            setToast({
-              hasError: true,
-              title: `Filter could not be applied!`,
-              message: `Please refresh the page and try again.`
-            });
-            setError(true);
-            setSearchedLoading(false);
-          } else {
-            setSearchedInventory(res);
-            setFilters(newFilters);
-            setDisplayedFilters(newFiltersToDisplay);
-            setSkippedSearch(prev => prev + batchSize);
-            setSearchedLoading(false);
-
-            if (res.length >= batchSize) {
-              setShouldFetchMore(true);
-            } else {
-              setShouldFetchMore(false);
-            }
-          }
-        });
-    }
-    return null;
   }
 
   /** Fetch inventory, top stats and hidden resources for a given custom view.
@@ -269,8 +199,21 @@ function useInventory() {
         setInventory,
         setSkipped
       });
+      getInventoryListFromAFilter({
+        router,
+        setSearchedLoading,
+        setStatsLoading,
+        setToast,
+        setError,
+        setInventoryStats,
+        batchSize,
+        setSearchedInventory,
+        setFilters,
+        setDisplayedFilters,
+        setSkippedSearch,
+        setShouldFetchMore
+      });
       getCustomViewInventoryListAndStats();
-      getInventoryListFromAFilter();
     }
   }, [router.query, views]);
